Add pipeline for KITS-19

+++ src/preprocessing/delete_imgs_with_no_annotations.py
@@ -2,11 +2,13 @@
 import glob
 import os
 
+import cv2
+import numpy as np
 from sklearn.base import TransformerMixin
 from tqdm import tqdm
 
 
-class DeleteImgsWithoutMasks(TransformerMixin):
+class DeleteImgsWithNoAnnotations(TransformerMixin):
     """Delete images without masks."""
 
     def __init__(
@@ -33,21 +35,34 @@ class DeleteImgsWithoutMasks(TransformerMixin):
         mask_paths = glob.glob(
             f"{os.path.join(root_path, self.mask_folder_name)}/**/*.png", recursive=True
         )
-        mask_names = [os.path.basename(mask) for mask in mask_paths]
+        self.mask_names = [os.path.basename(mask) for mask in mask_paths]
         print("Deleting images without masks...")
-        for img_path in tqdm(X):
-            img_name = os.path.basename(img_path)
-            if img_name not in mask_names:
-                self.delete_imgs_without_masks(img_path)
 
+        for img_path in tqdm(X):
+            self.delete_imgs_with_no_annotations(img_path, root_path)
+
+        # Create new list of paths after the deletion
         root_path = os.path.dirname(X[0])
         new_paths = glob.glob(f"{root_path}/*.png", recursive=True)
         return new_paths
 
-    def delete_imgs_without_masks(self, img_path: str) -> None:
-        """Delete images without masks.
+    def delete_imgs_with_no_annotations(self, img_path: str, root_path: str) -> None:
+        """Delete images without any annotations i.e. labels or masks.
 
         Args:
             img_path (str): Path to the image.
+            root_path (str): Path to the root folder.
         """
-        os.remove(img_path)
+        img_name = os.path.basename(img_path)
+        mask_path = os.path.join(root_path, self.mask_folder_name, img_name)
+        no_label = False
+        if "-" not in img_name:
+            no_label = True
+        if img_name not in self.mask_names:
+            if no_label:
+                os.remove(img_path)
+        # If there is a mask but it is empty
+        elif np.unique(cv2.imread(mask_path)).shape[0] == 1:
+            if no_label:
+                os.remove(img_path)
+                os.remove(mask_path)
